\name{x100}
\alias{x100}

\title{Simulated data for demonstrating blockTools functionality}
\description{Simulated data for demonstrating \code{blockTools} functionality.}
\usage{data(x100)}

\format{A dataframe with 100 rows and 6 columns.  Columns \code{id} and
  \code{id2} are unit and subunit identifying variables, \code{b1} and
  \code{b2} are blocking variables, \code{g} identifies which of three
  groups each unit is in (``a", ``b", or ``c"), and \code{ig} is an 
ignored
  variable. 
}
  
\author{Ryan T. Moore}

\examples{
data(x100)
}

\keyword{datasets}

