% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x100.R
\docType{data}
\name{x100}
\alias{x100}
\title{Simulated data for demonstrating blockTools functionality}
\format{
A dataframe with 100 rows and 6 columns.
\describe{
  \item{id}{Identifies units.}
  \item{id2}{Identifies subunits.}
  \item{b1}{Numeric blocking variable 1.}
  \item{b2}{Numeric blocking variable 2.}
  \item{g}{Character variable to identify which of three groups each unit is 
  in (\dQuote{a}, \dQuote{b}, or \dQuote{c}).}
  \item{ig}{Numeric variable that can be ignored.}
}
}
\usage{
data(x100)
}
\description{
Simulated data for demonstrating \code{blockTools} functionality.
}
\examples{
data(x100)
head(x100)

}
\author{
Ryan T. Moore
}
\keyword{datasets}
